Organized wallet definitions.
