--- src/connectors/index.ts
@@ -1,1 +1,0 @@
-export * from "./walletconnect-keplr"
