+++ src/utils/chainInfo.ts
@@ -1428,7 +1428,9 @@ export const getChainInfo = async (
     // Check overrides for chain info.
     overrides?.find((info) => info.chainId === chainId) ||
     // Use embedded map as fallback.
-    ChainInfoMap[chainId]
+    chainId in ChainInfoMap
+      ? ChainInfoMap[chainId as keyof typeof ChainInfoMap]
+      : undefined
 
   if (!chainInfo) {
     const availableChainIds = [
