+++ src/utils/index.ts
@@ -1,3 +1,2 @@
 export * from "./chainInfo"
 export * from "./getConnectedWalletInfo"
-export * from "./wallets"
