+++ src/wallets/index.ts
@@ -1,0 +1,5 @@
+import { Wallet } from "../types"
+import { wallets as keplrWallets } from "./keplr"
+import { wallets as leapWallets } from "./leap"
+
+export const WALLETS: Wallet[] = [...keplrWallets, ...leapWallets]
