+++ src/wallets/keplr/index.ts
@@ -1,0 +1,5 @@
+import { Wallet } from "../../types"
+import { KeplrExtensionWallet } from "./extension"
+import { KeplrMobileWallet } from "./mobile"
+
+export const wallets: Wallet[] = [KeplrExtensionWallet, KeplrMobileWallet]
