+++ src/wallets/leap/index.ts
@@ -1,0 +1,4 @@
+import { Wallet } from "../../types"
+import { LeapExtensionWallet } from "./extension"
+
+export const wallets: Wallet[] = [LeapExtensionWallet]
