+++ tsconfig.json
@@ -26,8 +26,7 @@
     "skipLibCheck": true,
     "sourceMap": true,
     "strict": true,
-    "strictNullChecks": true,
-    "suppressImplicitAnyIndexErrors": true
+    "strictNullChecks": true
   },
   "include": ["./src/**/*"]
 }
